# docs/swagger.yaml
basePath: /api
definitions:
  github_com_p4ndabk_base-project-go_internal_apierror.Body:
    properties:
      error:
        $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Detail'
    type: object
  github_com_p4ndabk_base-project-go_internal_apierror.Detail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_user.User:
    properties:
      active:
        type: boolean
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  internal_user.createRequest:
    properties:
      active:
        type: boolean
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_user.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_user.loginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/internal_user.User'
    type: object
  internal_user.updateRequest:
    properties:
      active:
        type: boolean
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    type: object
info:
  contact: {}
  description: Base Gin + GORM API.
  title: base-project-go API
  version: "1.0"
paths:
  /health:
    get:
      description: Pings the database to confirm the API and DB are reachable
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_user.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
      summary: Login
      tags:
      - auth
  /me:
    get:
      description: Reference example for protecting a route with AuthRequired.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
      security:
      - BearerAuth: []
      summary: Get the authenticated user
      tags:
      - auth
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_user.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_user.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
      summary: Create a user
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
      summary: Delete a user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
      summary: Get a user by id
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_user.updateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_p4ndabk_base-project-go_internal_apierror.Body'
      summary: Update a user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"